AOC 2023 Day 2 Typescript
This was a bit more regular expressions and more map and reduce.
Defaulting the input into the reduce for the colors was helpful.
There was a parsing problem where I was getting an undefined for the
last line in the test. There shouldn't be, yet worth looking at for
further learning.