AOC 2023 Day 3 Typescript
This took more parsing effort than I thought it would. The second stage
went quick as I was lucky to have done the correct parsing.

The major change was that I could not assume a * was the only valid
symbol for part 1. That meant having an extra function to do the work.
Maybe I could have done something clever like passing a function to the
core of the parts, yet that would be overkill. Changing the structure
a bit made the difference.