AOC 2023 Day 4 Typescript
The parsing was tried with a regular expression that was easier to do as
a split to pull out the parts. Afterward, part 1 went fast because I saw
that it was an exponent to get the count of winning cards.

Part 2 was more involved to try to understand the problem. It took more
and more reading to put it together. It wasn't just winning card, though
all cards found.

This was the opposite of prior days where it was about the parsing and
less about the items.