AOC 2023 Day 9 Typescript
The second part was trivial once the first part was done. I failed a
few times on the first part trying to get the result. I finally tracked
it down through the help of this video
https://www.youtube.com/watch?v=8qkH68u4tK0

I took a bit to realize the author posted their code, thanks @felerticia,
and that they did a neat trick in the video to not make the result match
what they showed for the input.

After determining that and checking my answer against his, I then submitted
to get the first star.

The error was a typical Javascript false error.  That I would not have had
if I had been in Javascript and not Typescript.
```
  const lastEntry = row.at(-1);
  if (!lastEntry || zeroCheck(row)) {
    return 0;
  }
```
Typescript complained that lower in the code lastEntry could be undefined
so I put in the `!lastEntry` as a check. This needed to be changed to be
`lastEntry == null` (I favor the coercion for undefined and null). Once
that was in place, if the last entry was zero I would not stop early.

That magic row produce a second to last row of -1 values.

All though it was most annoying to find, I tip my hat Eric Wastl for the
puzzle subtleties.