AOC 2023 Day 10 Typescript - Part 1
Fumbled my way through this one. I started with the approach of
trying to look at where I had come from in my moves. I then tried
to use a visited list. That made it worse.

Went back to returning where I came from after seeing that was the
needed approach from others. Got it to work and submitted.

Took a look at part 2. I should have remembered there is a formula
for calculating the points that are inside of a polygon. Found the
formula and code for it. That code is in this check in though not
used.