AOC 2023 Day 11 Typescript
This day I was able to go with the naive approach for part 1 and make
it work. After that it became a challenge of getting the math correct
to get the expansion of the universe correct.

After several pencil and paper attempts I landed on the equation of
```
row = row + (rowExpansions * expansion - rowExpansions);
```
I would have preferred to have made the expansion a parameter in my
tests to have multiple expansions. The math of the expansion was
the hard part as the code was simple after that.