AOC 2023 Day 12 Typescript - Part 1
This took lots of different attempts and several days. I knew what to
do just not how to do it. I only have the first part mostly right as
it can be a faster implementation.

I had to look to find how to recursively generate the bits of a number.
I knew that the working and damaged cells were the same as a 1 and a
zero in a binary number.

In order to do part 2 I will need to use memoization (dynamic
programming) to get to the answer. Even without looking at the problem.
That will also be a rework as I need to think in sub-problems and
concatenate the results of the sub-problems to give the answer to the
current problem.

Example:
??? is built of the first ? being either . or # and the result of the
?? being broken out. This will probably be a storage of
..# 1   #.# 1,1   .#. 1  #.. 1  ##. 2  .## 2