AOC 2023 Day 13 Typescript - Part 1
This was ugly. I thought I understood the problem and at simple level
I did. Then I struggled to see what was going on. I could get much of
the entry though struggled with others.

The idea of transposing the puzzle to do the second check made it
much easier to work with. Getting an algorithm to do the transpose
didn't work trying to do it myself and then I moved to finding one
online. Then I moved into the boundary conditions.

I am uncertain as to what in my changes of moving to a function of
validMirror did once I had that code correct. There was a basic error
that triggered too early or too late that caused problems that I
eventually went for help on.