AOC 2023 Day 14 Typescript - Part 1
This was an easier puzzle to solve than some of the earlier ones.
It is not entirely immutable though could be shifted to be that
way.

The part that took the most effort was understanding the AOC runner
running in PowerShell as I have shifted over to working in Windows
recently. A larger screen would have shown compile errors. Previously
the transpiling was not output and didn't clutter the console. Maybe
it was output and I just didn't see.

No matter. This is a good test of setting up the system for
development.